main-new created

+++ ch11/edgeimpulse_ex/edge-impulse-sdk/CMSIS/DSP/Source/BasicMathFunctions/arm_clip_f16.c
@@ -5,8 +5,10 @@
  * Title:        arm_clip_f16.c
  * Description:  Floating-point vector addition
  *
+ * $Date:        23 April 2021
+ * $Revision:    V1.9.0
  *
- * Target Processor: Cortex-M cores
+ * Target Processor: Cortex-M and Cortex-A cores
  * -------------------------------------------------------------------- */
 /*
  * Copyright (C) 2010-2021 ARM Limited or its affiliates. All rights reserved.
@@ -123,9 +125,9 @@ void arm_clip_f16(const float16_t * pSrc,
 {
     for (uint32_t i = 0; i < numSamples; i++)
     {                                        
-        if (pSrc[i] > high)                  
+        if ((_Float16)pSrc[i] > (_Float16)high)                  
             pDst[i] = high;                  
-        else if (pSrc[i] < low)              
+        else if ((_Float16)pSrc[i] < (_Float16)low)              
             pDst[i] = low;                   
         else                                 
             pDst[i] = pSrc[i];
